requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
flake8
python-dotenv>=0.5.1

numpy==1.18.1
matplotlib==3.1.2
sklearn
tensorflow==2.1.0
Keras==2.3.1
opencv-python==4.1.2.30
git+https://github.com/qubvel/efficientnet
git+https://github.com/qubvel/classification_models.git
git+https://github.com/qubvel/segmentation_models
git+https://github.com/albu/albumentations
tta-wrapper
jupyter
